requirements.txt
discord.py==2.3.2
stripe==5.4.0
aiohttp==3.8.5
python-dotenv==1.0.0
